WIP: Add inter-note link auto-complete
Adds an auto-complete provider to use Monaco's built-in functionality for
searching notes when linking. This patch presents an alternative approach
from creating a custom dialog.

Due to the way that the auto-complete always returns results even if none
are wanted it may be best to look for a custom dialog.